refactoring: extract logic to storage_pattern.py
